Move the S3 utility method deleteBucket to the AWS client utilities class
